refactor: init refactor
