+++ src/components/Game/BirdsList/BirdsItem/BirdItem.jsx
