+++ src/components/Game/EndGame/EndGame.jsx
